feat: cleanup lvm resources on delete
This commit cleans up the LVM resources created by the vgmanager
when the LVMCluster is deleted. All PVCs provisioned by topolvm
must be deleted before the LVMCluster is deleted. The LVMCluster
controller will wait until there are no existing LogicalVolumes before
it proceeds to remove the resources.
When the LVMCluster is deleted, the LVMCluster controller deletes the
LVMVolumeGroup resource. The resource is not removed immediately because
of the finalizer. The vg-manager deletes the vg on the host and updates
the LVMVolumeGroupNodeStatus, which triggers the LVMCluster
reconciliation to update the status. When the VG is removed from the
LVMCluster.status, the LVMCluster controller removes the finalizer on
the LVMVolumeGroup.

Signed-off-by: N Balachandran <nibalach@redhat.com>