fix: add ns labels for PSA
The PSA changes in OCP 4.12 prevent pods from starting as ns are
labeled restricted by default. This commit sets the label
pod-security.kubernetes.io/enforce: "privileged"
on the namespaces created by the e2e tests so pods can run.
This also renames a file to conform to the naming convention.

Signed-off-by: N Balachandran <nibalach@redhat.com>