build: move roles,clusterroles inside the CSV
having ClusterRoleBinding in the bundle prevents us to install a
operator in any namespace. As subjects[0].namespace of
ClusterRoleBinding is not populated via OLM if it is a independent
manifests inside a bundle.

operator-sdk allows the roles, rolebinding, clusterroles,
clusterrolebinding and serviceaccounts to be moved (using flag
--extra-service-accounts) to the CSV instead of being a independent
manifests inside a bundle. Moving them to CSV allows us to install
a operator in any namespace.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>