chore: refactor SCC creation and deletion in LVM Operator
This commit refactors the SecurityContextConstraint (SCC) creation and deletion in the LVM Operator. Previously, the check to determine if the SCC should be created or deleted was performed on every reconcile loop. This resulted in unnecessary API calls on non-Openshift clusters.

With this commit, the check is performed only once, during operator start-up and the information is stored in the LVMClusterReconciler structure. Therefore, the SCC resource is only manipulated if the operator is running on an Openshift cluster, saving resources on other cluster types.

This refactor makes the operator more efficient and potentially reduces the response time of reconcile loop when running on non-Openshift clusters.