fix: resource sync fail -> progressing
When the resource syncrhonization fails (e.g. because of an irrepairable error due to cluster state or due to the CSI node not being synchronized) we should move the status to progressing. That is because all of the resource sync problems are either resolvable automatically (after a certain time), or can be remediated by the user through changes in the cluster.

There is an edge that if we configure the pods wrong in the reconcile, then we have a permanent failure with state Progressing. However this is still technically correct since the Reconciler will keep trying to apply the resources (as opposed to perceived Failed which users would expect to not conclude with a retry)

Signed-off-by: Jakob Möller <jmoller@redhat.com>