symlink resolution: add caching
Add new object for symlink resolution with caching
to reduce symlink resolution amount within reconcile loop

Also, introduce a new type called DevicePath to contain user
provided device path. This is done in order to create
a standard way of working with raw(unresolved) device
path and avoid errors when it's not resolved while
it's expected to be.